@@ -2,14 +2,15 @@ allprojects {
     group 'enjoy-algorithm'
     version '1.0-SNAPSHOT'
 }
+apply from: rootProject.file('gradle/versions.gradle')
 
 subprojects {
     apply plugin: 'project-report'
     apply plugin: 'java'
     apply plugin: 'idea'
 
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_12
+    targetCompatibility = JavaVersion.VERSION_12
 
     repositories {
         mavenLocal()
@@ -19,7 +20,7 @@ subprojects {
     }
 
     dependencies {
-        //compile group: 'org.projectlombok', name:'lombok', version: "1.16.20"
+        compile group: 'org.projectlombok', name: 'lombok', version: "${lombokVersion}"
         compile group: 'com.google.code.findbugs', name: 'jsr305', version: '3.0.2'
         compile group: 'com.google.guava', name: 'guava', version: '21.0'
         compile group: 'org.jooq', name: 'jool', version: '0.9.12'
@@ -29,7 +30,12 @@ subprojects {
         compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.25'
         compile group: 'org.slf4j', name: 'jcl-over-slf4j', version: '1.7.25'
         testCompile group: 'org.mockito', name: 'mockito-core', version: '2.13.0'
-        testCompile group: 'junit', name: 'junit', version: '4.12'
+        testCompile group: 'org.junit.jupiter', name: 'junit-jupiter-params', version: "${junitJupiterVersion}"
+        testRuntime group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: "${junitJupiterVersion}"
+        testRuntime group: 'org.junit.platform', name: 'junit-platform-launcher', version: "${junitPlatformVersion}"
+        testRuntime group: 'org.junit.platform', name: 'junit-platform-engine', version: "${junitPlatformVersion}"
+        testRuntime group: 'org.junit.platform', name: 'junit-platform-commons', version: "${junitPlatformVersion}"
+        annotationProcessor group: 'org.projectlombok', name: 'lombok', version: "${lombokVersion}"
     }
 }
 buildscript {
